Assignment3/requirements.txt
# requirements.txt

fastapi==0.95.2
uvicorn==0.22.0
sqlmodel==0.0.8
sqlalchemy==1.4.41
pydantic==1.10.9
python-dotenv==1.0.0
pytest==7.4.0

